JSON-RPC site added.
In order to support more secure upload of videos, the server is now
required to provide clients with shared access signatures for uploading
videos so that the Azure storage account key does not need to be
distributed with the client application.  A JSON-RPC site has been added
with a single method - 'marathon.request_video_upload_url' - that
clients can call to retrieve a url with embedded shared access signature
to upload videos to.  This method requires that the user be
authenticated and takes a single argument: the guid of the video to be
uploaded.

This method